# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a documentação da API para o desafio.",
        "title": "Financial Exchange Challenge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/instruments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instruments"
                ],
                "summary": "Lista instrumentos financeiros",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base Asset",
                        "name": "base_asset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Quote Asset",
                        "name": "quote_asset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.InstrumentListDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo instrumento financeiro",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instruments"
                ],
                "summary": "Cria um novo instrumento",
                "parameters": [
                    {
                        "description": "Instrument",
                        "name": "instrument",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.CreateInstrumentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.CreateInstrumentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/instruments/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instruments"
                ],
                "summary": "Busca um instrumento por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instrument ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.InstrumentDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instruments"
                ],
                "summary": "Atualiza um instrumento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instrument ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Instrument",
                        "name": "instrument",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.CreateInstrumentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.InstrumentDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instruments"
                ],
                "summary": "Deleta um instrumento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instrument ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.CreateInstrumentRequest": {
            "type": "object",
            "required": [
                "base_asset",
                "quote_asset"
            ],
            "properties": {
                "base_asset": {
                    "type": "string"
                },
                "quote_asset": {
                    "type": "string"
                }
            }
        },
        "github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.CreateInstrumentResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.InstrumentDTO": {
            "type": "object",
            "properties": {
                "base_asset": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "quote_asset": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_mthpedrosa_financial-exchange-challenge_internal_instrument_domain_dto.InstrumentListDTO": {
            "type": "object",
            "properties": {
                "base_asset": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "quote_asset": {
                    "type": "string"
                }
            }
        },
...
    }
}